rename: components 폴더 삭제
